+++ src/foobara/ai/open_ai_api.rb
@@ -5,6 +5,16 @@ module Foobara
   module Ai
     module OpenAiApi
       foobara_domain!
+
+      class << self
+        def api_token=(token)
+          BaseCommand.api_token = token
+        end
+
+        def api_token
+          BaseCommand.api_token
+        end
+      end
     end
   end
 end