+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## [0.0.7] - 2025-02-17
+## [0.0.8] - 2025-02-17
 
 - Add support for some newly added fields
 - Make sure our models don't break when OpenAI adds new attributes to their responses
   before we have a chance to update them
 - Implement ListModels
+- Provide a way to set api token on project module
 
 ## [0.0.6] - 2025-02-16